pyproject.toml
[project]
name = "tensordb"
version = "1.0.0"
description = "Serverless database using Zarr, Fsspec and Xarray."
authors = [
    {name = "Joseph Gonzalez", email = "josephgonowak97@gmail.com"},
]
dependencies = [
    "pandas>=2.0.0",
    "xarray[accel]>=2023.0.0",
    "numpy>=1.26.0",
    "dask[complete]>=2023.0.0",
    "zarr>=2.0.0",
    "orjson>=3.0.0",
    "pydantic>=2.0.0",
    "more-itertools>=10.0.0",
    "loguru>=0.7.0",
    "numbagg>=0.8.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
